requirements.txt
tqdm
pyyaml
torch==2.4.1
numpy==2.1.1
pydantic==2.9.0
torchaudio==2.4.1
tensorboard>=2.4.1
torchvision==0.19.1
torch-tb-profiler>=0.2.1
opencv-python==4.10.0.84